EditHeightChart(NotFinished)

+++ Domain/Height.cs
@@ -6,7 +6,7 @@ namespace Domain
     {
         public Guid heightId { get; set; }
 
-        public double myHeight { get; set; }
+        public float myHeight { get; set; }
         public Patient patient { get; set; }
 
         public DateTime date { get; set;}
